pyproject.toml
[tool.poetry]
name = "dd2nx"
version = "0.1.1"
description = "Python library for converting the dd package's BDD to Multigraph in networkx."
readme="README.md"
repository = "https://github.com/mvcisback/dd2nx"
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
dd = "^0.5.4"
attrs = "^19.3"
funcy = "^1.13"
networkx = "^2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.7"
pytest-flake8 = "^1.0"
pytest-sugar = "^0.9.2"
pytest-xdist = "^1.29"
codecov = "^2.0"
pydot = "^1.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
